+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-2774] Structure types cache 
+
 ## [4.0.0] 2020-07-01
 
 ### Changed